Tests/LibWeb/Text/expected/DOM/Element-setAttribute-capitalization.txt

<root><div xmlns="http://www.w3.org/1999/xhtml" FoO="bar"></div></root>
